+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 
+### Added
+
+* Add new `LogoutRedirector` class
+
+### Changed
+
+* **[BC Break]** Most `LogoutSuccessHandler` logic moved to `LogoutRedirector`
+* **[BC Break]** Config key `handlers` renamed to `redirectors`
 
 ## [0.0.2] - 2020-04-15